requirements.txt
python-telegram-bot==20.4
pymongo>=4.0.1
python-dotenv>=0.19.2